requirements.txt
fastapi
requests
pydantic
uvicorn
firecrawl-py
